pyproject.toml
[project]
name = "pythonexecutiontool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.49.0",
    "black>=25.1.0",
    "boto3>=1.37.18",
    "isort>=6.0.1",
    "python-dotenv>=1.0.1",
    "semantic-kernel>=1.25.0",
    "tavily-python>=0.5.1",
]
